Refactoring code Fix error not table when trying
