+++ bot/bot.py
@@ -7,4 +7,3 @@ bot = telebot.TeleBot(TG_TOKEN)
 @bot.message_handler(content_types=['text'])
 def text_handler(chat_id, text):
     bot.send_message(chat_id, text)
-
